Added musl_math software library Re-organising cpu register banks (WIP) Re-organising zynqmp and SOCs tomls Changed and fixed bit rotten board handling Redid ident to require (in most games) 3 seperates, also both name and Definition Ident useable from project files Implemented hardware max instances/max cores and address offsets Fixed to new way ports and pin matches CPU folders named after cpu type not name Refactored Containers Added hidden attribute to CPP core header Refactor DDR structure Added IoDefinitionType Support fraction frame rates in zynqmp display port Made DistanceMatrix print more verbose (hiding unconnected chips) Added isIsolated to DistanceMatrix to test for unconnected chips Added FetchBootData to pmu_monitor Added board as command line parameters to overlord Added prefabs support Changed how pingroups are defined (in a table in the board object) refactored os functions in pmu_monitor Added board ident to software macro system Write register lists without register banks that have been used Added graph output to Report (outputs a graphviz dot file of the project) Added Xilinx software library for using the MIT libraries

+++ src/main/resources/catalogs/socs/zynqps8/software/libraries/display_port/include/zynqps8/display_port/display.hpp
@@ -318,7 +318,7 @@ struct VideoTiming {
 	uint16_t vTotal;
 	bool hSyncPolarity;
 	bool vSyncPolarity;
-	uint16_t frameRateHz;
+	float frameRateHz;
 };
 
 enum class MixerFunction {
